+++ srlc/website/urls.py
@@ -1,5 +1,4 @@
 import environ
-from debug_toolbar.toolbar import debug_toolbar_urls
 from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib import admin
@@ -86,6 +85,8 @@ urlpatterns = [
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
 
 if settings.DEBUG:
+    from debug_toolbar.toolbar import debug_toolbar_urls
+
     urlpatterns = urlpatterns + debug_toolbar_urls()
 
 handler404 = "srl.static_views.page_not_found"