+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v3.4.1
+###### July 18, 2025
+*   Added a new API field that displays the number of players within a game's subcategory.
+
+* * *
+
 ### v3.4
 ###### July 17, 2025